[depthMap] CUDA: Rename file OneTC
