+++ src/aliceVision/depthMap/cuda/planeSweeping/plane_sweeping_cuda.hpp
@@ -9,7 +9,7 @@
 #include <aliceVision/depthMap/SgmParams.hpp>
 #include <aliceVision/depthMap/RefineParams.hpp>
 #include <aliceVision/depthMap/cuda/commonStructures.hpp>
-#include <aliceVision/depthMap/cuda/tcinfo.hpp>
+#include <aliceVision/depthMap/cuda/OneTC.hpp>
 
 namespace aliceVision {
 namespace depthMap {
